Add `on-init` action
Sometimes you want to start the an upload as soon as the component is
rendered, and so you need to fire the `send` action immediately without
waiting for further events initiated by the user.

To do this, we add the `on-init` action which receives the request model
as its argument, and so any of the actions can be invoked there.